Conflicts:
	projects/openrct2.vcxproj.filters

+++ src/audio.c
@@ -577,15 +577,15 @@ MMRESULT mmio_open(const char* filename, HMMIO* hmmio, HGLOBAL* hmem, LPMMCKINFO
 				//strcpy(audio_info.var_0, "\x01");
 				hmem = 0;
 			label11:
-				hmemold2 = GlobalAlloc(0, hmem + 18);
+				hmemold2 = GlobalAlloc(0, (SIZE_T)(hmem + 18));
 				*hmemold = hmemold2;
 				if (!hmemold2) {
 					result = 57344;
 					goto label20;
 				}
 				memcpy(hmemold2, &waveformat, 16);
-				*((uint16*)*hmemold + 8) = hmem;
-				if (!hmem || mmioRead(hmmio1, (char*)*hmemold + 18, hmem) == hmem) {
+				*((uint16*)*hmemold + 8) = (uint16)hmem;
+				if (!hmem || mmioRead(hmmio1, (char*)*hmemold + 18, (LONG)hmem) == (LONG)hmem) {
 					result = mmioAscend(hmmio1, &mmckinfo1, 0);
 					if (!result) {
 						goto label24;
